requirements.txt
fastapi
mongoengine
passlib
pydantic
pydantic[email]
PyJWT
pymongo
python-dotenv
python-multipart
zxcvbn
blinker
uvicorn
pytest
pytest-mock
pytest-asyncio
locust
